backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
celery[redis]==5.4.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pydantic==2.10.2
pydantic-settings==2.6.1
alembic==1.14.0
reportlab==4.2.5
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.28.0